Split out OrderRow on Exchange and Account versions, differentiated by "dashboard" variable, cleaned up unused references by each component, etc.
