+++ app/components/Exchange/MyOpenOrders.jsx
@@ -1,6 +1,5 @@
 import React from "react";
 import PropTypes from "prop-types";
-import {Link} from "react-router-dom";
 import counterpart from "counterpart";
 import Ps from "perfect-scrollbar";
 import OpenSettleOrders from "./OpenSettleOrders";
@@ -13,10 +12,6 @@ import AssetName from "../Utility/AssetName";
 import Icon from "../Icon/Icon";
 import {ChainStore} from "bitsharesjs";
 import {LimitOrder, CallOrder} from "common/MarketClasses";
-import {EquivalentValueComponent} from "../Utility/EquivalentValueComponent";
-import {MarketPrice} from "../Utility/MarketPrice";
-import FormattedPrice from "../Utility/FormattedPrice";
-const leftAlign = {textAlign: "left"};
 const rightAlign = {textAlign: "right"};
 const centerAlign = {textAlign: "center"};
 import ReactTooltip from "react-tooltip";
@@ -99,7 +94,7 @@ TableHeader.defaultProps = {
     baseSymbol: null
 };
 
-class OrderRow extends React.Component {
+class ExchangeOrderRow extends React.Component {
     shouldComponentUpdate(nextProps) {
         return (
             nextProps.order.for_sale !== this.props.order.for_sale ||
@@ -116,7 +111,6 @@ class OrderRow extends React.Component {
             quote,
             order,
             showSymbols,
-            dashboard,
             isMyAccount,
             settings
         } = this.props;
@@ -133,11 +127,8 @@ class OrderRow extends React.Component {
         ) : null;
         let valueSymbol = showSymbols ? " " + base.get("symbol") : null;
         let amountSymbol = showSymbols ? " " + quote.get("symbol") : null;
-        let preferredUnit = settings ? settings.get("unit") : "1.3.0";
-        let quoteColor = !isBid ? "value negative" : "value positive";
-        let baseColor = isBid ? "value negative" : "value positive";
 
-        return !dashboard ? (
+        return (
             <tr key={order.id}>
                 <td className={tdClass} style={{paddingLeft: 10}}>
                     <PriceText
@@ -202,150 +193,11 @@ class OrderRow extends React.Component {
                     )}
                 </td>
             </tr>
-        ) : (
-            <tr key={order.id} className="clickable">
-                {isMyAccount ? (
-                    <td className="text-center">
-                        {isCall ? null : (
-                            <span
-                                style={{marginRight: 0}}
-                                className="order-cancel"
-                            >
-                                <input
-                                    type="checkbox"
-                                    className="orderCancel"
-                                    onChange={this.props.onCheckCancel}
-                                />
-                            </span>
-                        )}
-                    </td>
-                ) : null}
-                <td>
-                    <Link
-                        to={`/market/${quote.get("symbol")}_${base.get(
-                            "symbol"
-                        )}`}
-                    >
-                        <Icon
-                            name="trade"
-                            title="icons.trade.trade"
-                            className="icon-14px"
-                        />
-                    </Link>
-                </td>
-                <td style={leftAlign}>#{order.id.substring(4)}</td>
-                <td colSpan="4" style={leftAlign} onClick={this.props.onFlip}>
-                    {isBid ? (
-                        <Translate
-                            content="exchange.buy_description"
-                            baseAsset={utils.format_number(
-                                order[
-                                    isBid ? "amountToReceive" : "amountForSale"
-                                ]().getAmount({real: true}),
-                                base.get("precision"),
-                                false
-                            )}
-                            quoteAsset={utils.format_number(
-                                order[
-                                    isBid ? "amountForSale" : "amountToReceive"
-                                ]().getAmount({real: true}),
-                                quote.get("precision"),
-                                false
-                            )}
-                            baseName={
-                                <AssetName
-                                    noTip
-                                    customClass={quoteColor}
-                                    name={quote.get("symbol")}
-                                />
-                            }
-                            quoteName={
-                                <AssetName
-                                    noTip
-                                    customClass={baseColor}
-                                    name={base.get("symbol")}
-                                />
-                            }
-                        />
-                    ) : (
-                        <Translate
-                            content="exchange.sell_description"
-                            baseAsset={utils.format_number(
-                                order[
-                                    isBid ? "amountToReceive" : "amountForSale"
-                                ]().getAmount({real: true}),
-                                base.get("precision"),
-                                false
-                            )}
-                            quoteAsset={utils.format_number(
-                                order[
-                                    isBid ? "amountForSale" : "amountToReceive"
-                                ]().getAmount({real: true}),
-                                quote.get("precision"),
-                                false
-                            )}
-                            baseName={
-                                <AssetName
-                                    noTip
-                                    customClass={quoteColor}
-                                    name={quote.get("symbol")}
-                                />
-                            }
-                            quoteName={
-                                <AssetName
-                                    noTip
-                                    customClass={baseColor}
-                                    name={base.get("symbol")}
-                                />
-                            }
-                        />
-                    )}
-                </td>
-                <td style={leftAlign} onClick={this.props.onFlip}>
-                    <FormattedPrice
-                        base_amount={order.sellPrice().base.amount}
-                        base_asset={order.sellPrice().base.asset_id}
-                        quote_amount={order.sellPrice().quote.amount}
-                        quote_asset={order.sellPrice().quote.asset_id}
-                        force_direction={base.get("symbol")}
-                        hide_symbols
-                    />
-                </td>
-                <td style={leftAlign} onClick={this.props.onFlip}>
-                    {isBid ? (
-                        <MarketPrice
-                            base={base.get("id")}
-                            quote={quote.get("id")}
-                            force_direction={base.get("symbol")}
-                            hide_symbols
-                            hide_asset
-                        />
-                    ) : (
-                        <MarketPrice
-                            base={base.get("id")}
-                            quote={quote.get("id")}
-                            force_direction={base.get("symbol")}
-                            hide_symbols
-                            hide_asset
-                        />
-                    )}
-                </td>
-                <td style={{textAlign: "right"}} onClick={this.props.onFlip}>
-                    <EquivalentValueComponent
-                        hide_asset
-                        amount={order.amountForSale().getAmount()}
-                        fromAsset={order.amountForSale().asset_id}
-                        noDecimals={true}
-                        toAsset={preferredUnit}
-                    />{" "}
-                    <AssetName name={preferredUnit} />
-                </td>
-            </tr>
         );
     }
 }
 
-OrderRow.defaultProps = {
+ExchangeOrderRow.defaultProps = {
     showSymbols: false
 };
 
@@ -572,7 +424,7 @@ class MyOpenOrders extends React.Component {
                 .map(order => {
                     let price = order.getPrice();
                     return (
-                        <OrderRow
+                        <ExchangeOrderRow
                             price={price}
                             key={order.id}
                             order={order}
@@ -593,7 +445,7 @@ class MyOpenOrders extends React.Component {
                 .map(order => {
                     let price = order.getPrice();
                     return (
-                        <OrderRow
+                        <ExchangeOrderRow
                             price={price}
                             key={order.id}
                             order={order}
@@ -795,4 +647,4 @@ MyOpenOrders.propTypes = {
     baseSymbol: PropTypes.string.isRequired
 };
 
-export {OrderRow, TableHeader, MyOpenOrders};
+export {TableHeader, MyOpenOrders};
